Compilable with runtime symbol lookup error

+++ Simulation/Pathfinding/Interfaces/Graph.cpp
@@ -49,9 +49,9 @@ float Graph::heuristicCost(std::shared_ptr<Vertex> from, std::shared_ptr<Vertex>
             this->heuristicCosts[i][e->getEndVertex()->getId()] = e->getCost();
         }
     }
-    for (int k = 0; k < size; ++k){
-    for (int i = 0; i < size; ++i){
-    for (int j = 0; j < size; ++j){
+    for (size_t k = 0; k < size; ++k){
+    for (size_t i = 0; i < size; ++i){
+    for (size_t j = 0; j < size; ++j){
         float directCost = this->heuristicCosts[i][j];
         float indirectCost = this->heuristicCosts[i][k] + this->heuristicCosts[k][j];
         if (directCost > indirectCost)
